# Reduce wasteful  SystemCatalog  and Statistics procedure invocation