This commit ensures that we only call  `@SystemCatalog` and `@Statistics` once per `*Conn`