@gernest 2017-07-11
Benchmark comparison against master

```
benchcmp old.bench new.bench
benchmark                                                       old ns/op     new ns/op     delta
BenchmarkSerializeArgs-4                                        3843          3549          -7.65%
BenchmarkDeserializeResponse-4                                  1236          1032          -16.50%
BenchmarkDeserializeRows-4                                      349577        328636        -5.99%
BenchmarkDeserializeResult-4                                    1240          1135          -8.47%
BenchmarkExec-4                                                 2257583       724954        -67.89%
BenchmarkPrepareStmtExec-4                                      1582527       707875        -55.27%
BenchmarkQuery-4                                                6081415       4128942       -32.11%
BenchmarkPrepareStmtQuery-4                                     6740969       4787809       -28.97%
BenchmarkHashinater_getHashedPartitionForParameter_int32-4      251           235           -6.37%
BenchmarkHashinater_getHashedPartitionForParameter_int64-4      255           254           -0.39%
BenchmarkHashinater_getHashedPartitionForParameter_String-4     146           139           -4.79%
BenchmarkHashinater_getHashedPartitionForParameter_Bytes-4      312           311           -0.32%

benchmark                                                       old allocs     new allocs     delta
BenchmarkSerializeArgs-4                                        6              6              +0.00%
BenchmarkDeserializeResponse-4                                  6              6              +0.00%
BenchmarkDeserializeRows-4                                      839            845            +0.72%
BenchmarkDeserializeResult-4                                    12             12             +0.00%
BenchmarkExec-4                                                 1649           47             -97.15%
BenchmarkPrepareStmtExec-4                                      1086           50             -95.40%
BenchmarkQuery-4                                                2990           1055           -64.72%
BenchmarkPrepareStmtQuery-4                                     2871           1169           -59.28%
BenchmarkHashinater_getHashedPartitionForParameter_int32-4      1              1              +0.00%
BenchmarkHashinater_getHashedPartitionForParameter_int64-4      1              1              +0.00%
BenchmarkHashinater_getHashedPartitionForParameter_String-4     1              1              +0.00%
BenchmarkHashinater_getHashedPartitionForParameter_Bytes-4      0              0              +0.00%

benchmark                                                       old bytes     new bytes     delta
BenchmarkSerializeArgs-4                                        48            48            +0.00%
BenchmarkDeserializeResponse-4                                  80            80            +0.00%
BenchmarkDeserializeRows-4                                      1238673       1248005       +0.75%
BenchmarkDeserializeResult-4                                    80            80            +0.00%
BenchmarkExec-4                                                 301844        1742          -99.42%
BenchmarkPrepareStmtExec-4                                      199230        1954          -99.02%
BenchmarkQuery-4                                                3372144       2908845       -13.74%
BenchmarkPrepareStmtQuery-4                                     3261649       3241273       -0.62%
BenchmarkHashinater_getHashedPartitionForParameter_int32-4      32            32            +0.00%
BenchmarkHashinater_getHashedPartitionForParameter_int64-4      32            32            +0.00%
BenchmarkHashinater_getHashedPartitionForParameter_String-4     16            16            +0.00%
BenchmarkHashinater_getHashedPartitionForParameter_Bytes-4      0             0             +0.00%

```